Support taking submodule urls from .gitmodules
Use url from .gitmodules for git submodule layers.

The srcuri parameter of git layers in OESTACK should be removed. If left
there, it will not be used (the url from .gitmodules will override it)
but a warning will be isued in case they are not the same.

Git urls are still supported in OESTACK srcuri parameters, so it can still
be used for initial manifest creation.

As .gitmodules support relative url's, it is now possible to setup an
entire OE-lite manifest with all submodule urls being relative to the
manifest url.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>